base

+++ Codeforces/codeforces_1A_theatre-square.cpp
@@ -3,15 +3,15 @@
 
 int main()
 {
-    int n, m, a, res;
+    long long int n, m, a, res;
     std::cin >> n;
     std::cin >> m;
     std::cin >> a;
     
-    int x = n / a;
+    long long int x = n / a;
     while(x * a < n) x++;
     
-    int y = m / a;
+    long long int y = m / a;
     while (y * a < m) y++;
     
     res = x * y;
